z3-lift-arguments pass
This pass is a little complex, conceptually.  If we want to resolve a
function `f` but that function is called inside a function defn,
we normally wouldn't be able to resolve it because the argument isn't
statically resolvable. This pass makes the result of the call a new
argument and forces the caller to pass it in, making the whole thing
statically resolvable.

For example, consider the code:

    (define-fun f ((x Int)) Int x)
    (assert (= (f 1) 1))
    (define-fun g ((x Int)) Int (f x))
    (assert (= (g 1) 1))

Suppose we intend to statically resolve `f`. We can't just resolve
inside the definition of `g` (because we don't know that `x` is `1`),
but we might also not want to inline `g`. So instead, we make `(f x)`
and argument to `g`, like so:

    (define-fun f ((x Int)) Int x)
    (assert (= (f 1) 1))
    (define-fun g ((x Int) (fx Int) Int fx)
    (assert (= (g 1 (f 1)) 1))

Now we can statically resolve `(f 1)`, so after resolution this code
will have no calls to `f`.