New reimplementations of selectors
Selectors are now fully concrete. They are compiled to equivalence
classes (of element/property pairs) in Racket, including sorting the
selectors by precedence.

There is also an "inequalities" structure that we can test an
equivalence class set against. The idea is to enumerate selectors until
we find one that avoids all known inequalities and either get a model
synthesized or get a new inequality.