Lights! Handlers! Actions!
Starts adding rudimentary scripts to Cassius.

Scripts are composed of "actions" and "handlers".

Actions are events; currently only the "click" event is supported. These
trigger on particular elements, and can be asserted to convert one
document into another.

Handlers are scripts. They are fixed to certain elements, triggered by
events on those elements, and execute a program in a very simplistic
language to change the page.

For now I've committed the basic interpreter for scripts.