New matcher algorithm
The new matcher algorithm correctly handles the many and ugly variations
ANON boxes. The algorithm is, I think, fully correct, and also handles
inline-block (for future proofing).

get_bench.js is still not correct when it comes to extracting boxes
correctly, so this commit may actually increase the error rate.