Tighter but not complete above-the-mark semantics
Sometimes a line of text (or in rarer cases a flow root) needs to be
laid out north of the `ez.mark`. In these cases, we do not have info
about where floats are, so we cannot lay them out properly. We used to
not bound the position at all; now we bound it to at least be within the
parent and at or below its "no-float" position.